refactor(cli, api): move formatters from cli to api
